Extract JS Errors method to module
Add test, that JS error is not empty
